[Profiler] move out of ./utils/debug to ./module/profiler
