+++ module/profiler/profiler.go
@@ -1,4 +1,4 @@
-package debug
+package profiler
 
 import (
 	"context"
@@ -34,7 +34,8 @@ type AutoProfiler struct {
 	duration time.Duration
 }
 
-func NewAutoProfiler(log zerolog.Logger, dir string, interval time.Duration, duration time.Duration, enabled bool) (*AutoProfiler, error) {
+// New creates a new AutoProfiler instance performing profiling every interval for duration.
+func New(log zerolog.Logger, dir string, interval time.Duration, duration time.Duration, enabled bool) (*AutoProfiler, error) {
 	SetProfilerEnabled(enabled)
 
 	err := os.MkdirAll(dir, os.ModePerm)
